requirements.txt
tqdm
numpy
opencv-python
sklearn
scikit-image
pytorch-lightning
segmentation-models-pytorch
pytorch-fid
tensorboard
torch-tb-profiler
ipykernel
albumentations
kornia